!XI From //ce/task_array: New DynArray classes, features, optimisations. (Approved by thomasw) !FRO (Common) DynArray improvements: Added replace and emplace functions. Refactored all insertion functions to take all possible initialization values, including default- and no-init. Now inits to T() by default, same as std::vector. EInit argument can specify eZeroInit or eNoInit. Removed _raw functions. Optimized for moving temp values into array. Use variadic templates and forwarding to simplify replace, insert, etc. Resizing optimized, reallocation now occurs in 2 parts, reducing data movement to minimum required. !T (Common) Array storage renaming, and reformatting of single-line inline functions. Storage bases now implement begin() and end(). Removed slow MEMSTAT macros. Tweaked some internal typedefs. Replaced several derived types with type aliases. !R (Common) DynArray: Implemented LocalDynArray. Optimised StaticDynArray and FixedDynArray. Removed ResizeStorage. !B (Common) DynArray handles move-construction and swapping more correctly. Array storage classes now have dealloc() instead of destructors, fixing move behavior. Fixed alignment in HeapAllocator::Array. Added alignment asserts. !B (Common) Added DynArray unit tests, to verify element construct/destruct/move behavior. Review: ID 60765 by Author: Scott Peter, Reviewer: Matthijs, Observer: Nicolas Schulz, Observer: filipe amim, Observer: Timur Davidenko, Observer: Nico Moss, Observer: Domagoj Saric
Copied from Perforce
 Change: 1448708

+++ Code/CryEngine/CryAction/RuntimeAreas.cpp
@@ -157,7 +157,7 @@ void CRuntimeAreaManager::CreateAreas()
 			pAreaProxy = static_cast<IEntityAreaProxy*>(pNewAreaEntity->CreateProxy(ENTITY_PROXY_AREA).get());
 			if (pAreaProxy)
 			{
-				size_t const nPointCount = cluster.boundary_points.size();
+				int const nPointCount = cluster.boundary_points.size();
 				DynArray<bool> abObstructSound(nPointCount + 2, false);
 				points.resize(nPointCount);
 				for (size_t j = 0; j < points.size(); ++j)
